+++ docs/_downloads/257ac46232d6997b8582c10d2b58439a/sim_valve.py
@@ -5,7 +5,7 @@
 Example of a pneumatic valve being simulated until threshold is met. 
 """
 
-from prog_models.models.pneumatic_valve import PneumaticValve
+from progpy.models.pneumatic_valve import PneumaticValve
 
 def run_example(): 
     # Create a model object