+++ docs/_downloads/6a1ef040b1b1816faac11de7b6788010/visualize.py
@@ -7,7 +7,7 @@ Example demonstrating the Visualization Module.
 
 
 import matplotlib.pyplot as plt
-from prog_models.models.thrown_object import ThrownObject
+from progpy.models.thrown_object import ThrownObject
 
 def run_example():
     print('Visualize Module Example')