+++ docs/_downloads/82813b0c0a64173923320c20994e0946/vectorized.py
@@ -5,7 +5,7 @@
 Example using simulate_to_threshold with vectorized states. In this example we are using the thrown_object model to simulate multiple thrown objects
 """
 
-from prog_models.models.thrown_object import ThrownObject
+from progpy.models.thrown_object import ThrownObject
 from numpy import array, all
 
 def run_example():