+++ docs/_downloads/da191f26e484dddead89fb142ee68a12/sim.py
@@ -5,9 +5,9 @@
 Example of a battery being simulated for a set period of time and then till threshold is met.
 """
 
-from prog_models.models import BatteryCircuit as Battery
+from progpy.models import BatteryCircuit as Battery
 # VVV Uncomment this to use Electro Chemistry Model VVV
-# from prog_models.models import BatteryElectroChem as Battery
+# from progpy.models import BatteryElectroChem as Battery
 
 def run_example(): 
     # Step 1: Create a model object