+++ docs/_downloads/dd90523c294730d38024b15ef5fe8bf6/custom_model.py
@@ -16,8 +16,8 @@ import numpy as np
 from tensorflow import keras
 from tensorflow.keras import layers
 
-from prog_models.data_models import LSTMStateTransitionModel
-from prog_models.models import BatteryElectroChemEOD
+from progpy.data_models import LSTMStateTransitionModel
+from progpy.models import BatteryElectroChemEOD
 
 def run_example():
     print('Generating data...')