+++ sphinx-config/auto_examples/lstm_model.py
@@ -12,8 +12,8 @@ Finally, we repeat the exercise with data from the more complex BatteryElectroCh
 import matplotlib.pyplot as plt
 import numpy as np
 
-from prog_models.data_models import LSTMStateTransitionModel
-from prog_models.models import ThrownObject, BatteryElectroChemEOD
+from progpy.data_models import LSTMStateTransitionModel
+from progpy.models import ThrownObject, BatteryElectroChemEOD
 
 def run_example():
     # -----------------------------------------------------